Story 9-8: code-review patches → done
10 patches applied per Foundation Rule #6 (Medium+ findings re-pass).

2 Auditor BLOCKERS resolved:
- AC4 macro LOC: rewrote loan_status_badge.html from 59 → 23 LOC,
  trimmed verbose doc-comments while preserving role-gate logic.
- AC13 / Rule #12: extracted 240 LOC of `volume_detail_*` render
  tests from catalog.rs to NEW src/routes/volume_detail_tests.rs
  per the 9-6 home_indicator_tests.rs precedent. catalog.rs back to
  2440 LOC. Pre-9-8 over-2000 condition deferred as GH issue.

8 Medium+ findings patched:
- AC8 anonymous-leak test was tautological — added NEW
  volume_detail_anonymous_with_populated_borrower_data_does_not_leak
  rendering the anonymous variant with FULLY populated borrower
  fields. Locks the macro role-gate as the layer-2 defense the SQL
  projection narrowing alone can't catch in unit tests.
- Macro role gate `role != "anonymous"` failed OPEN on unknown role
  strings. Switched to POSITIVE whitelist `role == "librarian" ||
  role == "admin"` (fail-closed). Added regression guard
  volume_detail_unknown_role_falls_back_to_anonymous_variant
  simulating a Display impl rename to "Anonymous" (capital A).
- AC10 4-cell macro matrix complete (the missing security-relevant
  cell — anonymous role + populated borrower data — is now covered).
- LIMIT 1 without ORDER BY non-deterministic — added
  `ORDER BY loaned_at DESC, id DESC LIMIT 1` to both LoanModel
  methods so anonymous + librarian renders agree on the row choice
  if the data-integrity defense ever triggers.
- Anonymous query `(NaiveDateTime,)` tuple destructure → switched
  to `query_scalar::<_, NaiveDateTime>` per spec Task 1. CAST AS
  DATETIME RETAINED (mid-fix CI catch — sqlx 0.8 dynamic queries
  don't auto-decode TIMESTAMP per CLAUDE.md gotcha).
- `view_borrower_aria` dead plumbing — dropped i18n key + struct
  field + macro param + handler t!() call. 6 lines removed across
  5 files.
- E2E V-code `Date.now() % 10000` collision risk — introduced
  uniqueVcode() helper combining per-test counter + Date.now()%100
  (same pattern as 9-9 review fix).
- E2E `clearCookies()` on same context — HTMX GET-cache risk.
  Switched to `browser.newContext()` for the anonymous nav phase
  so a librarian-cached HTML response can't satisfy the assertion.

8 findings deferred to GH Issues per Foundation Rule #11 (date
format hardcoded ISO; soft-degrade returns None hides loan; FR
label width may overflow w-32; E2E volume-ID brute-force scan
capped at 200; ActiveLoanWithBorrower struct convention-only
enforcement; macro 8 positional params; anonymous /volume/{id}
extra DB roundtrip per page-view; catalog.rs over 2000 LOC pre-9-8).

Final 890/890 tests green (734 lib + 156 integration — +13 vs 9-9
review snapshot: 9 model unit tests + 1 setup_gate + 2 new
defense-in-depth render tests + 1 unknown-role test); clippy clean;
templates audit + i18n EN/FR mirror + tsc on tests/e2e/ + CI flake
gate all green; .sqlx cache untouched.

Story Status: review → done. sprint-status.yaml synced.
