+++ templates/components/loan_status_badge.html
@@ -1,59 +1,23 @@
-{# LoanStatusBadge component (story 9-8 AC4).
+{# LoanStatusBadge (story 9-8 AC4). Two role variants share one amber
+   wrapper. Anonymous: "On loan since {date}" (no PII). Librarian/Admin:
+   "On loan to <link> since {date}". The handler builds 3 i18n strings
+   so the macro can render the link between prefix + suffix without
+   template-side string-splitting (CSP-clean).
 
-   Renders the role-aware loan-status row for the volume-detail page
-   (and any future caller — title-detail volume list, search volume
-   rows, etc.). Two render variants based on `role`:
-
-   - **Anonymous**: "On loan since {date}" — NO borrower name, NO
-     /borrower/ link. Privacy-sensitive per FR59 (the SQL query layer
-     also narrows the projection — see two-layer defense in
-     LoanModel::active_loan_summary_for_volume).
-
-   - **Librarian/Admin**: "On loan to <name link> since {date}" —
-     borrower name is rendered as a plain link to /borrower/{id}.
-     The handler builds 3 separate i18n strings (label_anonymous +
-     label_prefix + label_suffix); the macro renders the link
-     between the prefix and suffix. This avoids template-side
-     string-splitting on an interpolated label and keeps the macro
-     CSP-clean.
-
-   Defense-in-depth: even if a future caller passes role="librarian"
-   without populating borrower fields, the macro falls through to the
-   anonymous variant rather than panicking on a None unwrap.
-
-   CSP-clean — Tailwind utility classes only, no inline styles, no
-   event handlers. The amber palette mirrors the existing
-   "not_shelved" badge in volume_detail.html — locks the "consistent
-   unavailability cue" UX contract. #}
-{%- macro badge(role, label_field, label_anonymous, label_prefix, label_suffix, borrower_id, borrower_name, view_borrower_aria) -%}
-{# `view_borrower_aria` parameter retained for API stability — currently
-   unused because the link's visible text (the borrower's name) is
-   already a meaningful accessible name; adding an `aria-label` would
-   hijack that for screen readers AND break Playwright's role-by-name
-   lookup (CI catch on PR #124 — `getByRole("link", { name: borrower })`
-   found zero matches when `aria-label="View borrower profile"` was
-   present). The parameter stays so future callers can opt back in
-   with a per-call accessible-name override if needed. #}
+   Role gate uses POSITIVE whitelist (`is_privileged`) so an unknown
+   role string fails CLOSED — falls through to the anonymous variant.
+   Defense-in-depth: librarian/admin role with None borrower data also
+   falls through. CSP-clean — Tailwind utility classes only. #}
+{%- macro badge(role, label_field, label_anonymous, label_prefix, label_suffix, borrower_id, borrower_name) -%}
+{%- let is_privileged = role == "librarian" || role == "admin" -%}
 <div class="flex items-center gap-3">
     <span class="text-sm font-medium text-stone-500 dark:text-stone-400 w-32">{{ label_field }}</span>
     <span class="inline-flex items-center gap-1 px-2 py-0.5 rounded-full bg-amber-100 dark:bg-amber-900/30 text-amber-700 dark:text-amber-400 text-xs font-medium">
-        {%- if role != "anonymous" -%}
-            {%- if let Some(bid) = borrower_id -%}
-                {%- if let Some(bname) = borrower_name -%}
-{# Librarian/Admin variant — borrower name as link between prefix + suffix. #}
-                    <span>{{ label_prefix }}</span><a href="/borrower/{{ bid }}" class="text-indigo-600 dark:text-indigo-400 hover:underline">{{ bname }}</a><span>{{ label_suffix }}</span>
-                {%- else -%}
-{# Defense-in-depth: librarian role but no borrower_name — fall back to anonymous variant. #}
-                    {{ label_anonymous }}
-                {%- endif -%}
-            {%- else -%}
-{# Defense-in-depth: librarian role but no borrower_id — fall back to anonymous variant. #}
-                {{ label_anonymous }}
-            {%- endif -%}
-        {%- else -%}
-{# Anonymous variant — date only, NO borrower data. #}
-            {{ label_anonymous }}
-        {%- endif -%}
+        {%- if is_privileged -%}
+            {%- if let Some(bid) = borrower_id -%}{%- if let Some(bname) = borrower_name -%}
+<span>{{ label_prefix }}</span><a href="/borrower/{{ bid }}" class="text-indigo-600 dark:text-indigo-400 hover:underline">{{ bname }}</a><span>{{ label_suffix }}</span>
+            {%- else -%}{{ label_anonymous }}{%- endif -%}{%- else -%}{{ label_anonymous }}{%- endif -%}
+        {%- else -%}{{ label_anonymous }}{%- endif -%}
     </span>
 </div>
 {%- endmacro -%}
